+++ android/sdk/lolila_sdk/src/main/cpp/base/geometry/LineDistance.cpp
@@ -11,11 +11,11 @@
 
 
 bool LineDistance::isParallel(const Vector& V1,const Vector& V2){
-    const double V1V2 = Vector::dot(V1,V2);
+    const float V1V2 = Vector::dot(V1,V2);
     Matrix M(2,2);
-    M.set(0,0,Vector::dot(V1,V1));M.set(0,1,-1.0*V1V2);
-    M.set(1,0,V1V2);              M.set(1,1,-1.0*Vector::dot(V2,V2));
-    return FloatUtils::isEqual(M.determinant(),0.0);
+    M.set(0,0,Vector::dot(V1,V1));M.set(0,1,-1.0f*V1V2);
+    M.set(1,0,V1V2);              M.set(1,1,-1.0f*Vector::dot(V2,V2));
+    return FloatUtils::isEqual(M.determinant(),0.0f);
 }
 
 /**
@@ -24,7 +24,7 @@ bool LineDistance::isParallel(const Vector& V1,const Vector& V2){
  * 关键是求解线性方程组
  * 判断是否有解，主要是看detM是否为0，若为01，则两条直线平行
  */
-double LineDistance::calc(const Ray& l1,const Ray& l2){
+float LineDistance::calc(const Ray& l1,const Ray& l2){
     LogQueue::push("LineDistance::calc begin...\n");
 
     const Vector& V1 = l1.direction();
@@ -33,20 +33,20 @@ double LineDistance::calc(const Ray& l1,const Ray& l2){
     const Vector& V2 = l2.direction();
     LogQueue::push("V2 is:%s\n",V2.c_str());
 
-    const double V1V2 = Vector::dot(V1,V2);
+    const float V1V2 = Vector::dot(V1,V2);
     Matrix M(2,2);
-    M.set(0,0,Vector::dot(V1,V1));M.set(0,1,-1.0*V1V2);
-    M.set(1,0,V1V2);              M.set(1,1,-1.0*Vector::dot(V2,V2));
+    M.set(0,0,Vector::dot(V1,V1));M.set(0,1,-1.0f*V1V2);
+    M.set(1,0,V1V2);              M.set(1,1,-1.0f*Vector::dot(V2,V2));
     LogQueue::push("M is:%s",M.c_str());
 
-    const double detM = M.determinant();
+    const float detM = M.determinant();
     LogQueue::push("detM is:%.8lf\n",detM);
 
-    if( FloatUtils::isEqual(detM,0.0)){
+    if( FloatUtils::isEqual(detM,0.0f)){
         LogQueue::push("Ll and L2 is parallel!\n");
         const Vector pAtL2 = l2.pointAt(0);
         LogQueue::push("get first point on L2: %s\n",pAtL2.c_str());
-        const double result = l1.distance(pAtL2);
+        const float result = l1.distance(pAtL2);
         LogQueue::push("distance between L1 and L2 is:%g\n",result);
         LogQueue::push("LineDistance::calc end.\n");
         return result;
@@ -68,36 +68,36 @@ double LineDistance::calc(const Ray& l1,const Ray& l2){
     LogQueue::push("R is: %s",R.c_str());
 
 
-    const Matrix T = Matrix::multiply(Matrix::scalarMultiply(M.cofactors(),1.0/detM),R);
+    const Matrix T = Matrix::multiply(Matrix::scalarMultiply(M.cofactors(),1.0f/detM),R);
     LogQueue::push("T is: %s",T.c_str());
 
 
-    const double  t1 = T.get(0,0);
+    const float  t1 = T.get(0,0);
     const Vector p1 = l1.pointAt(t1);
     LogQueue::push("point at L1: %s\n",p1.c_str());
 
-    const double  t2 = T.get(1,0);
+    const float  t2 = T.get(1,0);
     const Vector p2 = l2.pointAt(t2);
     LogQueue::push("point at L2: %s\n",p2.c_str());
 
 
-    const double distance = FloatUtils::minus(Vector::dot(p1,p1) + Vector::dot(p2,p2), 2.0*(Vector::dot(p1,p2)));
-    const double  result = sqrt(distance);
+    const float distance = FloatUtils::minus(Vector::dot(p1,p1) + Vector::dot(p2,p2), 2.0f*(Vector::dot(p1,p2)));
+    const float  result = sqrtf(distance);
     LogQueue::push("distance between L1 and L2 is:%.8lf\n",result);
     //LogQueue::push("distance v2 between L1 and L2 is:%g\n",Vector::minus(p1,p2).length());
     LogQueue::push("LineDistance::calc end.\n");
     return result;
 }
 
-double LineDistance::calc(const Line& l1, const Line& l2){
+float LineDistance::calc(const Line& l1, const Line& l2){
     return calc(Line::line2ray(l1),Line::line2ray(l2));
 }
 
-double LineDistance::calc(const Line& l1, const Ray& l2){
+float LineDistance::calc(const Line& l1, const Ray& l2){
     return calc(Line::line2ray(l1),l2);
 }
 
-double LineDistance::calc(const Ray& l1,const Line& l2){
+float LineDistance::calc(const Ray& l1,const Line& l2){
     return calc(l1,Line::line2ray(l2));
 }
