+++ android/sdk/lolila_sdk/src/main/cpp/base/geometry/Plane.cpp
@@ -16,13 +16,13 @@ bool  Plane::pointLiesInPlane(const Plane &plane, const Vector &Q){
     const Vector& N = plane._n;
     const Vector& P = plane._p;
     const Vector Q_P = Vector::minus(Q,P);
-    return FloatUtils::isEqual(Vector::dot(N,Q_P),0);
+    return FloatUtils::isEqual(Vector::dot(N,Q_P),0.0f);
 }
 
 bool  Plane::pointLiesInPlane( const Vector& Q)const{
     return pointLiesInPlane(*this, Q);
 }
 
-double Plane::distanceFromOri()const{
-    return fabs(D())/N().length();
+float Plane::distanceFromOri()const{
+    return fabsf(D())/N().length();
 }
